+++ CHANGELOG.md
@@ -2,6 +2,10 @@
 
 All notable changes to the "cph-ng" extension will be documented in this file.
 
+## [Unreleased]
+
+- **feat**: enhance path handling for moved workspace in problem data
+
 ## 0.4.8
 
 - **feat**: drag and drop to webview